+++ tourcount/src/main/java/com/wmstein/tourcount/SettingsFragment.kt
@@ -9,10 +9,11 @@ import com.wmstein.tourcount.TourCountApplication.Companion.getPrefs
 
 /**
  * SettingsFragment
+ * Copyright 2016-2026 wmstein
  * Created by wmstein on 2020-04-17
  * last edited in Java on 2020-04-18,
  * converted to Kotlin on 2023-07-06,
- * last edited on 2025-12-31
+ * last edited on 2026-01-03
  */
 class SettingsFragment : PreferenceFragmentCompat() {
     private var prefs: SharedPreferences? = null
@@ -24,15 +25,23 @@ class SettingsFragment : PreferenceFragmentCompat() {
         // Get preferences
         prefs = getPrefs()
 
-        // Set proximity option visible if available in device
+        // Set proximity or ambien light option visible if available in device
         val prefProx: Boolean = prefs!!.getBoolean("enable_prox", false)
         val proxPref: ListPreference? = findPreference("pref_prox")
         proxPref?.isEnabled = prefProx
+        if (prefProx)
+            proxPref?.setIcon(R.drawable.ic_speaker_phone_black_48dp)
+        else
+            proxPref?.setIcon(R.drawable.ic_speaker_phone_gray_48dp)
 
         // Set vibrator option visible if available in device
-        val prefVib: Boolean = prefs!!.getBoolean("", false)
+        val prefVib: Boolean = prefs!!.getBoolean("enable_vib", false)
         val vibPref: SwitchPreferenceCompat? = findPreference("pref_button_vib")
         vibPref?.isEnabled = prefVib
+        if (prefVib)
+            vibPref?.setIcon(R.drawable.outline_vibration_48)
+        else
+            vibPref?.setIcon(R.drawable.outline_vibration_gray_48)
     }
 
 }