+++ tourcount/src/main/java/com/wmstein/tourcount/Utils.kt
@@ -3,19 +3,20 @@ package com.wmstein.tourcount
 import android.text.Html
 import android.text.Spanned
 
-/******************************************************
- * Class Utils cares for Android versions compatibility
- *   in ShowTextDialog fromHtml
+/***********************************************
+ * object Utils cares for
+ * - formatted Html in ShowTextDialog and Toasts
  *
- * Created by wistein on 2017-09-25,
+ * Copyright 2016-2026 wmstein
+ * Created by wmstein on 2017-09-25,
  * last modified in Java on 2018-06-13,
  * converted to Kotlin on 2024-09-30,
- * last edited on 2025-07-15.
+ * last edited on 2026-01-02.
  */
 internal object Utils {
     @JvmStatic
     fun fromHtml(source: String?): Spanned {
         return Html.fromHtml(source, Html.FROM_HTML_MODE_LEGACY)
     }
-
 }
+