+++ docs/History.md
@@ -1,9 +1,9 @@
 # History
 
-#### TourCount Copyright © 2016-2025, Wilhelm Stein, Bonn, Germany
+#### TourCount Copyright © 2016-2026, Wilhelm Stein, Bonn, Germany
 
-#### Version 3.7.0 (2025-12-31)
-- Code improvement: Gray-out preference item "Button Vibration" when not available in device
+#### Version 3.7.0 (2026-01-03)
+- Code improvement: Gray-out preference items "Button Vibration" and "Proximity Sensor" when not available in device
 - Bugfix for sound-related crash under Android versions 7.1 - 10
 
 #### Version 3.6.9 (2025-12-28)